backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.6
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.7
cycler==0.12.1
databricks-sdk==0.105.0
docker==7.1.0
fastapi==0.136.1
filelock==3.29.0
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.47
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==25.3.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huey==2.6.0
huggingface_hub==1.12.0
hyperframe==6.1.0
idna==3.13
ImageHash==4.3.2
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
mlflow==3.11.1
mlflow-skinny==3.11.1
mlflow-tracing==3.11.1
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
opentelemetry-api==1.41.1
opentelemetry-proto==1.41.1
opentelemetry-sdk==1.41.1
opentelemetry-semantic-conventions==0.62b1
packaging==26.2
pandas==2.3.3
pillow==12.2.0
portalocker==3.2.0
prettytable==3.17.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psycopg2-binary==2.9.12
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
pytz==2026.1.post1
PyWavelets==1.9.0
PyYAML==6.0.3
qdrant-client==1.17.1
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
skops==0.14.0
smmap==5.0.3
SQLAlchemy==2.0.49
sqlparse==0.5.5
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==5.6.2
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.46.0
wcwidth==0.6.0
Werkzeug==3.1.8
yarl==1.23.0
zipp==3.23.1
